feat: scripting updates
