--- lib/vm/langs/python/lib.py
@@ -1,7 +1,0 @@
-import js
-import json
-import pyodide
-import time
-
-def vm_log(data):
-  js.postMessage(json.dumps({"action":"log","payload":data}))
