+++ middleware.ts
@@ -22,7 +22,7 @@ function getLocale(request: NextRequest): string | undefined {
 export function middleware(request: NextRequest) {
   // Check if there is any supported locale in the pathname
   const pathname = request.nextUrl.pathname
-  if (pathname.startsWith('/vm')) {
+  if (['/vm', '/webworker'].some((p) => pathname.startsWith(p))) {
     return
   }
