+++ ui/lesson/ScriptingChallenge/Editor.tsx
@@ -1,0 +1,64 @@
+'use client'
+
+import MonacoEditor from '@monaco-editor/react'
+
+import { monacoOptions } from './config'
+import { monaco } from 'react-monaco-editor'
+
+export default function Editor({
+  language,
+  value,
+  onChange,
+  onValidate,
+}: {
+  language: string
+  value: string
+  onChange?: (value: string) => void
+  onValidate?: (value: monaco.editor.IMarker[]) => void
+}) {
+  const handleBeforeMount = (monaco) => {
+    monaco.editor.defineTheme('satoshi', {
+      base: 'vs-dark',
+      inherit: true,
+      rules: [],
+      colors: {
+        // 'editor.background': '#7B1810',
+      },
+    })
+
+    // monaco.languages.typescript.javascriptDefaults.addExtraLib(
+    //   [
+    //     'declare const Facts = {',
+    //     '    /**',
+    //     '     * Returns the next fact',
+    //     '     */',
+    //     '    static next():string',
+    //     '}',
+    //   ].join('\n'),
+    //   'filename/facts.d.ts'
+    // )
+  }
+
+  const handleMount = (editor, monaco) => {
+    monaco.editor.setTheme('satoshi')
+  }
+
+  return (
+    <div className="bg-[#432A43] font-mono text-sm text-white">
+      <MonacoEditor
+        width="calc(100vw / 2)"
+        height="calc(100vh - 71px - 48px - 40px - 160px)"
+        // theme="satoshi"
+        // defaultLanguage={language}
+        // defaultValue={code}
+        language={language}
+        value={value}
+        beforeMount={handleBeforeMount}
+        onMount={handleMount}
+        onChange={onChange}
+        onValidate={onValidate}
+        options={monacoOptions}
+      />
+    </div>
+  )
+}
