+++ ui/lesson/ScriptingChallenge/config.ts
@@ -1,0 +1,26 @@
+import { editor } from 'monaco-editor'
+
+export const monacoOptions: editor.IStandaloneEditorConstructionOptions = {
+  readOnly: false,
+  minimap: { enabled: false },
+  // overviewRulerLanes: 0,
+  // overviewRulerBorder: false,
+  // scrollbar: {
+  //   vertical: 'auto',
+  //   horizontal: 'hidden',
+  //   handleMouseWheel: true,
+  // },
+  scrollBeyondLastLine: false,
+  wordWrap: 'on',
+}
+
+export const languageMeta = {
+  javascript: {
+    value: 'javascript',
+    label: 'JavaScript',
+  },
+  python: {
+    value: 'python',
+    label: 'Python',
+  },
+}
